separating documents and chunks
I additionally renamed folder containing models for code readability from documents into MyModels
