+++ api/MyModels/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class DocumentsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'documents'
+    name = 'MyModels'
