+++ api/MyModels/chunks/schemas.py
@@ -14,12 +14,3 @@ class ChunkOut(Schema):
     chunk_idx: int
     start_char: int
     end_char: int
-
-class DocumentIn(Schema):
-    text: str
-    embedding: list[float]
-
-class DocumentOut(Schema):
-    id: int
-    text: str
-    embedding: list[float]
