+++ api/MyModels/documents/schemas.py
@@ -1,0 +1,10 @@
+from ninja import Schema
+
+class DocumentIn(Schema):
+    text: str
+    embedding: list[float]
+
+class DocumentOut(Schema):
+    id: int
+    text: str
+    embedding: list[float]
