+++ api/MyModels/documents/views.py
@@ -5,43 +5,10 @@ from django.http import HttpRequest
 from ninja import Router
 from ninja.pagination import LimitOffsetPagination, paginate
 
-from documents.models import Chunk, Document
-from documents.schemas import ChunkIn, ChunkOut, DocumentIn, DocumentOut
+from MyModels.models import Document
+from MyModels.documents.schemas import DocumentIn, DocumentOut
 
-chunk_router = Router()
 document_router = Router()
-
-@chunk_router.post("/chunk/", response={HTTPStatus.CREATED: ChunkOut})
-def create_chunk(request: HttpRequest, payload: ChunkIn):
-    chunk = Chunk(**payload.dict())
-    chunk.full_clean()
-    chunk.save()
-    return HTTPStatus.CREATED, chunk
-
-@chunk_router.get("/chunk/", response={HTTPStatus.OK: List[ChunkOut]})
-@paginate(LimitOffsetPagination)
-def list_chunks(request: HttpRequest):
-    return Chunk.objects.all()
-
-@chunk_router.get("/chunk/{id}/", response={HTTPStatus.OK: ChunkOut})
-def retrieve_chunk(request: HttpRequest, id: int):
-    chunk = Chunk.objects.get(id=id)    
-    return chunk
-
-@chunk_router.put("/chunk/{id}/", response={HTTPStatus.OK: ChunkOut})
-def update_chunk(request: HttpRequest, id: int, payload: ChunkIn):
-    chunk = Chunk.objects.get(id=id)
-    for attr, value in payload.dict(exclude_unset=True).items():
-        setattr(chunk, attr, value)
-    chunk.full_clean()
-    chunk.save()
-    return chunk
-
-@chunk_router.delete("/chunk/{id}/", response={HTTPStatus.OK: None})
-def delete_chunk(request: HttpRequest, id: int):
-    chunk = Chunk.objects.get(id=id)
-    chunk.delete()
-    return HTTPStatus.OK
 
 @document_router.post("/document/", response={HTTPStatus.CREATED: DocumentOut})
 def create_document(request: HttpRequest, payload: DocumentIn):
