+++ api/MyModels/urls.py
@@ -2,5 +2,5 @@ from django.urls import path
 from .api import api
 
 urlpatterns = [
-    path("api/", api.urls),
+    path("", api.urls),
 ]
