+++ api/api/settings.py
@@ -29,7 +29,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "documents",
+    "MyModels.apps.DocumentsConfig",
 ]
 
 MIDDLEWARE = [
